% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomplete_keys.R
\name{absent_keys}
\alias{absent_keys}
\title{Key combinations not present in input}
\usage{
absent_keys(df, key_cols = NULL, non_key_cols = NULL)
}
\arguments{
\item{df}{Dataframe to inspect for incomplete key combinations.}

\item{key_cols}{Character vector of key columns.}

\item{non_key_cols}{Character vector of non-key columns. You cannot provide both \code{key_cols} and \code{non_key_cols}.}
}
\value{
Dataframe.
}
\description{
This function returns a dataframe of key column combinations that are not present in \code{df}.
Thus if \code{df} has no key duplicates and \code{nrow(absent_keys(df))==0} then \code{df} is complete in its keys.
}
\details{
By default if neither \code{key_cols} or  \code{non_key_cols} are provided, all columns are interpreted as key columns.
}
\examples{
> a <- tibble(key_1 = c(1,1,2),
              key_2 = c(1,2,1))

> absent_keys(a)
# A tibble: 1 × 2
}
