% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_age_group_labels.R
\name{make_age_group_labels}
\alias{make_age_group_labels}
\title{Make character age-group labels from numeric age-group breaks.}
\usage{
make_age_group_labels(age_group_breaks)
}
\arguments{
\item{age_group_breaks}{Numeric vector of age group breaks}
}
\value{
Character vector
}
\description{
The age-group breaks are interpreted as open to the left and closed to the right.
See examples.
}
\examples{
> make_age_group_labels(c(0,5,13,Inf))
"0-4"    "5-12"   "13plus"
}
