% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_key_cols.R
\name{select_key_cols}
\alias{select_key_cols}
\title{Helper for selecting columns by inclusion or exclusion but not both.}
\usage{
select_key_cols(df, key_cols = NULL, non_key_cols = NULL)
}
\arguments{
\item{df}{Dataframe to select columns from.}

\item{key_cols}{Columns to select.}

\item{non_key_cols}{Columns to disclude. You cannot provide both \code{key_cols} and \code{non_key_cols}.}
}
\value{
Dataframe
}
\description{
Helper for selecting columns by inclusion or exclusion but not both.
}
\examples{
> a <- tibble(key_1 = numeric(),
              key_2 = numeric(),
              non_key = numeric())

> select_key_cols(a, key_cols = c("key_1", "key_2"))
# A tibble: 0 × 2

> select_key_cols(a, non_key_cols = "non_key")
# A tibble: 0 × 2
… with 2 variables: key_1 <dbl>, key_2 <dbl>

> select_key_cols(a, key_cols = c("key_1", "key_2"), non_key_cols = "non_key")
Error in select_key_cols(a, key_cols = c("key_1", "key_2"), non_key_cols = "non_key") :
  You can define which variables should be checked for completeness by setting 'key_cols'
  or which variables should be discluded by setting 'non_key_cols'.
  You cannot set both.
}
